packages/orchestrator/pyproject.toml
[project]
name = "ours-aaif-orchestrator"
version = "0.1.0"
description = "Orchestrator Agent for OURS AAIF platform"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
anthropic = [
    "langchain-anthropic>=0.1.0",
]
openai = [
    "langchain-openai>=0.1.0",
]
google = [
    "langchain-google-genai>=1.0.0",
]
all = [
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ours_aaif_orchestrator"]
